[refactor] move each class in a separate file
