+++ src/spdx3/model/software/package.py
@@ -15,7 +15,7 @@ from spdx3.model.creation_information import CreationInformation
 from common.typing.type_checks import check_types_and_set_values
 
 from common.typing.dataclass_with_properties import dataclass_with_properties
-from spdx3.model.element import Artifact
+from spdx3.model.artifact import Artifact
 from spdx3.model.software.software_purpose import SoftwarePurpose
